Reorganize folders and remove sass from SlidingPanel
