+++ components/index.scss
@@ -23,8 +23,8 @@
 @import "./radioButton/radioButton.scss";
 @import "./rating/rating.scss";
 @import "./selectionBlock/selectionBlock.css";
-@import "./slidingPanel/slidingPanel.scss";
-@import "./slidingPanel/slidingPanelHeader.scss";
+@import "./slidingPanel/slidingPanel.css";
+@import "./slidingPanel/slidingPanelHeader/slidingPanelHeader.css";
 @import "./spinner/spinner.scss";
 @import "./tabs/tab.css";
 @import "./tabs/tabs.css";
