+++ components/slidingPanel/slidingPanel.css
@@ -8,9 +8,10 @@
   top: 0;
   z-index: var(--tx-sliding-panel-overlay-z-index);
 
-  &.ui-sliding-panel-overlay_hidden {
-    display: none;
-  }
+}
+
+.ui-sliding-panel-overlay.ui-sliding-panel-overlay_hidden {
+  display: none;
 }
 
 .ui-sliding-panel {
@@ -26,9 +27,10 @@
   position: var(--tx-sliding-panel-panel-position);
   top: var(--tx-sliding-panel-panel-top);
 
-  &.ui-sliding-panel_active {
-    opacity: var(--tx-sliding-panel-panel-open-opacity);
-  }
+}
+
+.ui-sliding-panel.ui-sliding-panel_active {
+  opacity: var(--tx-sliding-panel-panel-open-opacity);
 }
 
 .ui-sliding-panel_right {
@@ -36,26 +38,20 @@
   left: var(--tx-sliding-panel-right-direction-left);
   right: var(--tx-sliding-panel-right-direction-right);
 
-  &.ui-sliding-panel_active  {
-    animation: sidePanelOpen 195ms ease-in-out 0s 1 both;
-  }
+}
+
+.ui-sliding-panel_right.ui-sliding-panel_active  {
+  animation: sidePanelOpen 195ms ease-in-out 0s 1 both;
 }
 
 .ui-sliding-panel_left {
   animation: sidePanelLeftClose 195ms ease-in-out 0s 1 both;
   left: var(--tx-sliding-panel-left-direction-left);
   right: var(--tx-sliding-panel-left-direction-right);
-
-  &.ui-sliding-panel_active  {
-    animation: sidePanelLeftOpen 195ms ease-in-out 0s 1 both;
-  }
 }
 
-.ui-sliding-panel__content {
-  flex: var(--tx-sliding-panel-content-flex);
-  overflow: var(--tx-sliding-panel-content-overflow);
-  overflow-y: var(--tx-sliding-panel-content-overflow-y);
-  padding: 20px;
+.ui-sliding-panel_left.ui-sliding-panel_active  {
+  animation: sidePanelLeftOpen 195ms ease-in-out 0s 1 both;
 }
 
 .ui-sliding-panel__footer {
